+++ crates/algorithms/ed25519/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+- [#1241](https://github.com/cryspen/libcrux/pull/1241): 
+    - Add `Ed25519KeyPair` type
+    - Add key-centric signature public APIs
+
 ## [0.0.4] (2025-11-05)
 
 ## [0.0.3] (2025-06-30)